refactor: 다이얼로그를 전역으로 관리 (#476)
* feat: 새로운 Dialog 추가

전역 Dialog Configs 설정 및 간단히 추가하는 훅 추가

* feat: 다이얼로그 호출방식 추가

* feat: 같은 키값의 Dialog 무시하도록

* refactor: useApi 에러 처리 기본값 새로운 다이얼로그로 변경

* refactor: 다이얼로그 활용방식 변경 useDialog -> useNewDialog

+++ src/component/superTag/SuperTagMergeAccordion.tsx
@@ -3,12 +3,12 @@ import { useGetTagsSuperTagIdSub } from "../../api/tags/useGetTagsSuperTagIdSub"
 import { useDeleteTagsSuper } from "../../api/tags/useDeleteTagsSuper";
 import { usePatchTagsBookInfoIdMerge } from "../../api/tags/usePatchTagsBookInfoIdMerge";
 import { Tag } from "../../type";
-import { useDialog } from "../../hook/useDialog";
 import Accordion from "../utils/Accordion";
 import SuperTagMergeSubTag from "./SuperTagMergeSubTag";
 import Droppable from "../utils/Droppable";
 import Image from "../utils/Image";
 import TrashIcon from "../../asset/img/trash.svg";
+import { useNewDialog } from "../../hook/useNewDialog";
 
 type Props = {
   bookInfoId: number;
@@ -17,22 +17,10 @@ type Props = {
 };
 
 const SuperTagMergeAccordion = ({ bookInfoId, tag, removeTag }: Props) => {
-  const { Dialog, setOpenTitleAndMessage, setConfig, setOpen, defaultConfig } =
-    useDialog();
   const { subTagList, toggleOpened, addSubTag, removeSubTag } =
-    useGetTagsSuperTagIdSub({
-      tagId: tag.id,
-      setOpenTitleAndMessage,
-    });
-  const { deleteTag } = useDeleteTagsSuper({
-    removeTag,
-    setOpenTitleAndMessage,
-  });
-
-  const { setParams } = usePatchTagsBookInfoIdMerge({
-    bookInfoId,
-    setOpenTitleAndMessage,
-  });
+    useGetTagsSuperTagIdSub({ tagId: tag.id });
+  const { deleteTag } = useDeleteTagsSuper({ removeTag });
+  const { setParams } = usePatchTagsBookInfoIdMerge({ bookInfoId });
 
   const addNewListAndMergeIfMoved: DragEventHandler = e => {
     const stringifiedTag = e.dataTransfer.getData("text/plain");
@@ -43,26 +31,16 @@ const SuperTagMergeAccordion = ({ bookInfoId, tag, removeTag }: Props) => {
     }
   };
 
+  const { addConfirmDialog } = useNewDialog();
   const getConfirmThenRemove: MouseEventHandler = e => {
     e.preventDefault();
     e.stopPropagation();
-    setConfig({
-      ...defaultConfig,
-      title: `태그를 삭제하겠습니까?`,
-      message: `태그 내용 : ${tag.content}`,
-      numberOfButtons: 2,
-      firstButton: {
-        ...defaultConfig.firstButton,
-        onClick: () => {
-          deleteTag(tag.id);
-        },
-      },
-    });
-    setOpen();
+    addConfirmDialog("삭제확인", "태그를 삭제하겠습니까?", "", () =>
+      deleteTag(tag.id),
+    );
   };
   return (
     <div className="super-tag__accordion__wrapper">
-      <Dialog />
       <Accordion
         summaryButtonClassName="super-tag__accordion__summary"
         summaryUI={
